Refactor all startup code into `@storybook/core` also
