+++ app/angular/package.json
@@ -49,7 +49,6 @@
     "react": "^16.0.0",
     "react-dev-utils": "^5.0.0",
     "react-dom": "^16.0.0",
-    "redux": "^3.6.0",
     "sass-loader": "^6.0.7",
     "ts-loader": "^3.5.0",
     "uglifyjs-webpack-plugin": "^1.2.4",
