+++ app/polymer/package.json
@@ -51,7 +51,6 @@
     "react": "^16.0.0",
     "react-dev-utils": "^5.0.0",
     "react-dom": "^16.0.0",
-    "redux": "^3.7.2",
     "uglifyjs-webpack-plugin": "^1.2.4",
     "util-deprecate": "^1.0.2",
     "webpack": "^3.11.0",
