+++ app/react/package.json
@@ -57,7 +57,6 @@
     "markdown-loader": "^2.0.2",
     "prop-types": "^15.6.1",
     "react-dev-utils": "^5.0.0",
-    "redux": "^3.7.2",
     "uglifyjs-webpack-plugin": "^1.2.4",
     "util-deprecate": "^1.0.2",
     "webpack": "^3.11.0",
