+++ app/vue/package.json
@@ -49,7 +49,6 @@
     "react": "^16.2.0",
     "react-dev-utils": "^5.0.0",
     "react-dom": "^16.2.0",
-    "redux": "^3.7.2",
     "uglifyjs-webpack-plugin": "^1.2.4",
     "util-deprecate": "^1.0.2",
     "webpack": "^3.11.0",
