+++ lib/core/src/client/preview/index.js
@@ -1,3 +1,4 @@
+import start from './start';
 import * as Actions from './actions';
 import ClientApi from './client_api';
 import ConfigApi from './config_api';
@@ -5,4 +6,4 @@ import StoryStore from './story_store';
 import reducer from './reducer';
 import syncUrlWithStore from './syncUrlWithStore';
 
-export default { Actions, ClientApi, ConfigApi, StoryStore, reducer, syncUrlWithStore };
+export default { start, Actions, ClientApi, ConfigApi, StoryStore, reducer, syncUrlWithStore };
