+++ lib/core/src/client/preview/start.js
@@ -1,0 +1,65 @@
+import { createStore } from 'redux';
+import addons from '@storybook/addons';
+import { navigator, window } from 'global';
+import createChannel from '@storybook/channel-postmessage';
+import { handleKeyboardShortcuts } from '@storybook/ui/dist/libs/key_events';
+
+import StoryStore from './story_store';
+import ClientApi from './client_api';
+import ConfigApi from './config_api';
+import reducer from './reducer';
+import * as Actions from './actions';
+import syncUrlWithStore from './syncUrlWithStore';
+
+export default function start(render, { decorateStory } = {}) {
+  // check whether we're running on node/browser
+  const isBrowser =
+    navigator &&
+    navigator.userAgent &&
+    navigator.userAgent !== 'storyshots' &&
+    !(navigator.userAgent.indexOf('Node.js') > -1) &&
+    !(navigator.userAgent.indexOf('jsdom') > -1);
+
+  const storyStore = new StoryStore();
+  const reduxStore = createStore(reducer);
+  const context = { storyStore, reduxStore, decorateStory };
+
+  const clientApi = new ClientApi(context);
+
+  let channel;
+  if (isBrowser) {
+    // setup preview channel
+    channel = createChannel({ page: 'preview' });
+    channel.on('setCurrentStory', data => {
+      reduxStore.dispatch(Actions.selectStory(data.kind, data.story));
+    });
+    addons.setChannel(channel);
+    Object.assign(context, { channel });
+
+    syncUrlWithStore(reduxStore);
+
+    // Handle keyboard shortcuts
+    window.onkeydown = handleKeyboardShortcuts(channel);
+  }
+
+  // Provide access to external scripts if `window` is defined.
+  // NOTE this is different to isBrowser, primarily for the JSDOM use case
+  if (typeof window !== 'undefined') {
+    window.__STORYBOOK_CLIENT_API__ = clientApi;
+    window.__STORYBOOK_ADDONS_CHANNEL__ = channel; // may not be defined
+  }
+
+  const { clearDecorators } = clientApi;
+  const configApi = new ConfigApi({ clearDecorators, ...context });
+
+  // initialize the UI
+  const renderUI = () => {
+    if (isBrowser) {
+      render(context);
+    }
+  };
+
+  reduxStore.subscribe(renderUI);
+
+  return { context, clientApi, configApi, forceReRender: () => render(context, true) };
+}
